+++ benefits/src/sdk/sdkconfiguration.py
@@ -2,6 +2,7 @@
 
 
 import requests as requests_http
+from ._hooks import SDKHooks
 from .utils import utils
 from .utils.retries import RetryConfig
 from dataclasses import dataclass
@@ -25,10 +26,11 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '4.0.2'
-    gen_version: str = '2.262.2'
-    user_agent: str = 'speakeasy-sdk/python 4.0.2 2.262.2 1.0.0 wingspan_benefits'
+    sdk_version: str = '4.1.0'
+    gen_version: str = '2.277.0'
+    user_agent: str = 'speakeasy-sdk/python 4.1.0 2.277.0 1.0.0 wingspan_benefits'
     retry_config: RetryConfig = None
+    _hooks: SDKHooks = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]:
         if self.server_url:
@@ -37,3 +39,7 @@ class SDKConfiguration:
             self.server_idx = 0
 
         return SERVERS[self.server_idx], {}
+
+
+    def get_hooks(self) -> SDKHooks:
+        return self._hooks