+++ benefits/RELEASES.md
@@ -130,3 +130,13 @@ Based on:
 - [python v4.0.2] benefits
 ### Releases
 - [PyPI v4.0.2] https://pypi.org/project/wingspan_benefits/4.0.2 - benefits
+
+## 2024-03-04 01:27:30
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.200.0 (2.277.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.1.0] benefits
+### Releases
+- [PyPI v4.1.0] https://pypi.org/project/wingspan_benefits/4.1.0 - benefits